fix for failing benchmark tests
